# docs/openapi_spec.yaml
openapi: 3.1.0
info:
  title: Trendit API
  description: Comprehensive Reddit Data Collection and Analysis Platform
  version: 1.0.0
paths:
...
  /api/query/posts:
    post:
      tags:
      - query
      summary: Query Posts
      description: "Advanced post query with comprehensive filtering options.\n\n\
        Supports complex queries like:\n- Posts from multiple subreddits with keyword\
        \ filtering\n- Score and engagement thresholds  \n- Author inclusion/exclusion\n\
        - Content type filtering\n- Date range filtering"
      operationId: query_posts_api_query_posts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api__query__PostQueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/comments:
    post:
      tags:
      - query
      summary: Query Comments
      description: 'Advanced comment query with comprehensive filtering options.


        Supports queries like:

        - Comments from specific posts or subreddits

        - Score and depth filtering

        - Author filtering

        - Content filtering'
      operationId: query_comments_api_query_comments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api__query__CommentQueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/users:
    post:
      tags:
      - query
      summary: Query Users
      description: 'Advanced user analysis and filtering.


        Supports queries like:

        - User profiles by username

        - Active users in subreddits

        - Users by karma thresholds

        - Account age filtering'
      operationId: query_users_api_query_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserQueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/posts/form:
    post:
      tags:
      - query
      summary: Query Posts Form
      description: Query posts using form parameters (easier to use in Swagger UI
        than JSON)
      operationId: query_posts_form_api_query_posts_form_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_query_posts_form_api_query_posts_form_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/query/posts/simple:
    get:
      tags:
      - query
      summary: Simple Post Query
      description: Simple post query via GET parameters
      operationId: simple_post_query_api_query_posts_simple_get
      parameters:
      - name: subreddits
        in: query
        required: true
        schema:
          type: string
          description: Comma-separated subreddit names
          title: Subreddits
        description: Comma-separated subreddit names
      - name: keywords
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Comma-separated keywords
          title: Keywords
        description: Comma-separated keywords
      - name: min_score
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Minimum score
          title: Min Score
        description: Minimum score
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          description: Maximum results
          default: 50
          title: Limit
        description: Maximum results
      - name: sort_type
        in: query
        required: false
        schema:
          type: string
          description: 'Sort type: hot, new, top, rising, controversial'
          default: hot
          title: Sort Type
        description: 'Sort type: hot, new, top, rising, controversial'
      - name: time_filter
        in: query
        required: false
        schema:
          type: string
          description: 'Time filter: hour, day, week, month, year, all'
          default: week
          title: Time Filter
        description: 'Time filter: hour, day, week, month, year, all'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueryResponse:
      properties:
        query_type:
          type: string
          title: Query Type
        parameters:
          additionalProperties: true
          type: object
          title: Parameters
        results:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Results
        count:
          type: integer
          title: Count
        execution_time_ms:
          type: number
          title: Execution Time Ms
        reddit_api_calls:
          type: integer
          title: Reddit Api Calls
        filters_applied:
          items:
            type: string
          type: array
          title: Filters Applied
      type: object
      required:
      - query_type
      - parameters
      - results
      - count
      - execution_time_ms
      - reddit_api_calls
      - filters_applied
      title: QueryResponse
      description: Standard query response format
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    api__query__PostQueryRequest:
      properties:
        subreddits:
          items:
            type: string
          type: array
          title: Subreddits
          description: List of subreddit names
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
          description: Keywords to search for
        exclude_keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Exclude Keywords
          description: Keywords to exclude
        date_from:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
          description: Start date (ISO format)
        date_to:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
          description: End date (ISO format)
        min_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Score
          description: Minimum post score
        max_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Score
          description: Maximum post score
        min_upvote_ratio:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Upvote Ratio
          description: Minimum upvote ratio (0.0-1.0)
        max_upvote_ratio:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Upvote Ratio
          description: Maximum upvote ratio (0.0-1.0)
        min_comments:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Comments
          description: Minimum number of comments
        max_comments:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Comments
          description: Maximum number of comments
        include_authors:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Include Authors
          description: Include only these authors
        exclude_authors:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Exclude Authors
          description: Exclude these authors
        exclude_deleted:
          type: boolean
          title: Exclude Deleted
          description: Exclude deleted posts
          default: true
        exclude_removed:
          type: boolean
          title: Exclude Removed
          description: Exclude removed posts
          default: true
        content_types:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Content Types
          description: 'Content types: text, link, image, video'
        exclude_nsfw:
          type: boolean
          title: Exclude Nsfw
          description: Exclude NSFW content
          default: true
        exclude_spoilers:
          type: boolean
          title: Exclude Spoilers
          description: Exclude spoiler content
          default: true
        exclude_stickied:
          type: boolean
          title: Exclude Stickied
          description: Exclude stickied posts
          default: true
        sort_type:
          type: string
          title: Sort Type
          description: 'Sort type: hot, new, top, rising, controversial'
          default: hot
        time_filter:
          type: string
          title: Time Filter
          description: 'Time filter: hour, day, week, month, year, all'
          default: all
        limit:
          type: integer
          title: Limit
          description: Maximum results (1-1000)
          default: 100
        include_self_text:
          type: boolean
          title: Include Self Text
          description: Include post self text
          default: true
        include_awards:
          type: boolean
          title: Include Awards
          description: Include award information
          default: false
      type: object
      required:
      - subreddits
      title: PostQueryRequest
      description: Advanced post query parameters